@@ -1,26 +1,44 @@
 [build-system]
-requires = ["setuptools", "wheel"]
-build-backend = "setuptools.build_meta"
+requires = [
+    "hatchling",
+]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.sdist]
+only-include = [
+    "src/ospx",
+    "tests",
+    ".coveragerc",
+    ".editorconfig",
+    "pytest.ini",
+    "ruff.toml",
+    "uv.lock",
+]
+
+[tool.hatch.build.targets.wheel]
+packages = [
+    "src/ospx",
+]
 
 [project]
 name = "ospx"
 version = "0.2.14"
 description = "Extension package to farn, adding support to build OSP simulation cases using FMUs."
 readme = "README.md"
-requires-python = ">= 3.9"
-license = {file = "LICENSE"}
+requires-python = ">= 3.10"
+license = { file = "LICENSE" }
 authors = [
-    {name = "Frank Lumpitzsch", email = "frank.lumpitzsch@dnv.com"},
-    {name = "Claas Rostock", email = "claas.rostock@dnv.com"},
+    { name = "Frank Lumpitzsch", email = "frank.lumpitzsch@dnv.com" },
+    { name = "Claas Rostock", email = "claas.rostock@dnv.com" },
 ]
 maintainers = [
-    {name = "Claas Rostock", email = "claas.rostock@dnv.com"},
+    { name = "Claas Rostock", email = "claas.rostock@dnv.com" },
 ]
-keywords = []
+keywords = [
+]
 classifiers = [
     "Development Status :: 4 - Beta",
     "License :: OSI Approved :: MIT License",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
