@@ -1,28 +1,24 @@
 import logging
 import os
-from glob import glob
 from pathlib import Path
 from shutil import rmtree
 
 import pytest
-from pytest import LogCaptureFixture
 
 from ospx.utils.zip import add_file_content_to_zip
 
 
 @pytest.fixture(scope="package", autouse=True)
-def chdir():
+def chdir() -> None:
     os.chdir(Path(__file__).parent.absolute() / "test_dicts")
 
 
 @pytest.fixture(scope="package", autouse=True)
-def test_dir():
+def test_dir() -> Path:
     return Path(__file__).parent.absolute()
 
 
-output_dirs = [
-    "xyz",
-]
+output_dirs = []
 output_files = [
     "parsed*",
     "*.xml",
@@ -36,7 +32,7 @@ output_files = [
 
 
 @pytest.fixture(autouse=True)
-def default_setup_and_teardown(caplog: LogCaptureFixture):
+def default_setup_and_teardown():
     _remove_output_dirs_and_files()
     _create_test_fmu()
     yield
@@ -44,14 +40,14 @@ def default_setup_and_teardown(caplog: LogCaptureFixture):
     _remove_output_dirs_and_files()
 
 
-def _remove_output_dirs_and_files():
+def _remove_output_dirs_and_files() -> None:
     for folder in output_dirs:
         rmtree(folder, ignore_errors=True)
     for pattern in output_files:
-        for file in glob(pattern):
-            file = Path(file)
-            if not file.name.startswith("test_"):
-                file.unlink(missing_ok=True)
+        for file in Path.cwd().glob(pattern):
+            _file = Path(file)
+            if not _file.name.startswith("test_"):
+                _file.unlink(missing_ok=True)
 
 
 def _create_test_fmu():
@@ -73,11 +69,11 @@ def _remove_test_fmu():
 
 
 @pytest.fixture(autouse=True)
-def setup_logging(caplog: LogCaptureFixture):
-    caplog.set_level("WARNING")
+def setup_logging(caplog: pytest.LogCaptureFixture) -> None:
+    caplog.set_level("INFO")
     caplog.clear()
 
 
 @pytest.fixture(autouse=True)
-def logger():
+def logger() -> logging.Logger:
     return logging.getLogger()
